+++ tornado_battery/schema.py
@@ -1,10 +1,12 @@
-from tornado_battery.exception import ClientException
+from tornado.options import options
+from tornado_battery.exception import GeneralException, ClientException
 from marshmallow import Schema
 import functools
+import traceback
 import ujson
 
 
-def schema(query=None, form=None, json=None, reply=None):
+def schema(query=None, form=None, json=None, reply=None, error=None):
 
     def _load_query(handler, schema):
         data = {}
@@ -46,24 +48,40 @@ def schema(query=None, form=None, json=None, reply=None):
             if json and isinstance(json, Schema):
                 kwargs.update({'json': _load_json(handler, json)})
 
-            retval = await function(handler, *args, **kwargs)
-
-            if reply and isinstance(reply, Schema):
-                assert isinstance(retval, dict)
-                handler.set_header('Content-Type',
-                                   'application/json; charset=UTF-8')
-                data = reply.dump(retval['data']).data
-                return handler.finish(dict(code=retval.get('code', 0),
-                                           msg=retval.get('msg', ''),
-                                           data=data))
-            elif reply and isinstance(reply, bool):
-                # reply == True
-                assert isinstance(retval, dict)
-                handler.set_header('Content-Type',
-                                   'application/json; charset=UTF-8')
-                return handler.finish(retval)
-
-            return retval
+            try:
+                retval = await function(handler, *args, **kwargs)
+            except GeneralException as e:
+                if callable(error):
+                    handler.finish(error(handler, e))
+                elif error and isinstance(error, bool):
+                    errmsg = dict(code=e.error_code, msg=e.message)
+                    if hasattr(options, 'debug') and options.debug:
+                        errmsg['traceback'] = traceback.format_exc()
+                    handler.set_status(e.http_status_code)
+                    handler.finish(errmsg)
+                else:
+                    raise
+            except Exception as e:
+                if callable(error):
+                    handler.finish(error(handler, e))
+                else:
+                    raise
+            else:
+                if reply and isinstance(reply, Schema):
+                    assert isinstance(retval, dict)
+                    handler.set_header('Content-Type',
+                                       'application/json; charset=UTF-8')
+                    data = reply.dump(retval['data']).data
+                    return handler.finish(dict(code=retval.get('code', 0),
+                                               msg=retval.get('msg', ''),
+                                               data=data))
+                elif reply and isinstance(reply, bool):
+                    # reply == True
+                    assert isinstance(retval, dict)
+                    handler.set_header('Content-Type',
+                                       'application/json; charset=UTF-8')
+                    return handler.finish(retval)
+                return retval
         return f
 
     return wrapper